refactor: 뷰 생성자 주입
