+++ src/main/kotlin/blackjack/Application.kt
@@ -1,7 +1,9 @@
 package blackjack
 
 import blackjack.controller.BlackJackController
+import blackjack.view.Console
 
 fun main() {
-    BlackJackController().start()
+    val console = Console()
+    BlackJackController(console, console).start()
 }
